Reorganize the code, closes #76
